Added CompanyUserResponse.java and CustomerUserResponse.java that combines Company and User or Customer and User respectively
Added UpdateUserRequest.java which takes all possible fields and depending on role updates the correct fields for the Company or Customer user

Fixed UserController.java endpoints so they now work as intended and return either a CompanyUserResponse or CustomerUserResponse depending on role. The Delete endpoint just returns a string.

Added service methods with business logic in UserService.java so that the endpoints now function correctly.

Moved the validateNotAlreadyRegistered method from ValidationService.java to UserService.java and removed the UserService injection inside ValidationService.java. We did this to avoid a circular dependency issue and make it so that ValidationService.java has no contact with the database.

Added @OnDelete(action = OnDeleteAction.CASCADE) over the User field in ConfirmationToken.java so that when a User is deleted, then all the ConfirmationTokens related to that User are removed