User Service logic implemented and Zuul User Id Validation (#41)
* Added CompanyUserResponse.java and CustomerUserResponse.java that combines Company and User or Customer and User respectively

Added UpdateUserRequest.java which takes all possible fields and depending on role updates the correct fields for the Company or Customer user

Fixed UserController.java endpoints so they now work as intended and return either a CompanyUserResponse or CustomerUserResponse depending on role. The Delete endpoint just returns a string.

Added service methods with business logic in UserService.java so that the endpoints now function correctly.

Moved the validateNotAlreadyRegistered method from ValidationService.java to UserService.java and removed the UserService injection inside ValidationService.java. We did this to avoid a circular dependency issue and make it so that ValidationService.java has no contact with the database.

Added @OnDelete(action = OnDeleteAction.CASCADE) over the User field in ConfirmationToken.java so that when a User is deleted, then all the ConfirmationTokens related to that User are removed

* Added UserSecurity.java component with a method hasUserId that checks weather the id in the JWT token matches the path variable

Added antMatcher in WebSecurityConfiguration.java that calls hasUserId method.

* Added a if statement in hasUserId to check if a token has been set and if not return false.

* New method for validating userId in Zuul using id from token in header

Co-authored-by: Elias Mehr <55871294+EliasMehr@users.noreply.github.com>

* Added validation that header is not null and starts with "Bearer " inside hasSameIdAsHeader method in UserSecurity.java
Changed permitAll() to anonymous() for "user/register/**" antMatcher in WebSecurityConfiguration.java

Co-authored-by: daniel-hughes-nackademin <55873880+daniel-hughes-nackademin@users.noreply.github.com>
Co-authored-by: Elias Mehr <55871294+EliasMehr@users.noreply.github.com>
Co-authored-by: Daniel Hughes <daniel.hughes.sweden@gmail.com>